+++ core/components/romanesco/docs/changelog.md
@@ -8,6 +8,7 @@ New features:
 - Add preview and status properties to all elements
 
 Fixes and improvements:
+- Skip own element properties in returnFirstHit snippet
 - Add option to adjust column size to content width in CB settings
 - Scroll back to top after using pagination in overviews (can be disabled)
 - Only display avatar image in publication elements if author page exists